pyproject.toml
[tool.poetry]
name = "ytdl_downloader"
version = "0.1.0"
description = "Tool to query a database for URLs to download, attempt to download and update the database for successful downloads"
authors = ["Kim van Wyk <kimv@sahomeloans.com>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.1.24"
black = "^20.8b1"
attrs = "^20.3.0"
python-dotenv = "^0.19.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
ptpython = "^3.0.13"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
